requirements.txt
    pandas
    scikit-learn
    tensorflow-macos  # Específico para Mac
    nltk
    spacy
    matplotlib
    seaborn
    


